E5 fix dtos input/output fuente dinamica

+++ src/main/java/ar/utn/frba/ddsi/cliente_liviano/models/repositories/dto/FuenteDinamicaHechoResponse.java
@@ -25,7 +25,7 @@ public class FuenteDinamicaHechoResponse {
     private String descripcion;
 
     @JsonProperty("categoria")
-    private Integer categoria;
+    private Long categoria;
 
     @JsonProperty("ubicacionLatitud")
     private double ubicacionLatitud;
